Добавил обработку ошибки PropertyValueException (поле не может быть равно null). Избавился от склееных импортов. Изменил тип id с Long на BigInteger. Добавил JavaDoc к классу DefaultHandler.

+++ antifraud/src/main/java/com/bank/antifraud/controller/SuspiciousAccountTransfersController.java
@@ -1,14 +1,19 @@
 package com.bank.antifraud.controller;
 
-import com.bank.antifraud.dto.AuditDto;
 import com.bank.antifraud.dto.SuspiciousAccountTransfersDto;
-import com.bank.antifraud.entity.AuditEntity;
 import com.bank.antifraud.entity.SuspiciousAccountTransfersEntity;
-import com.bank.antifraud.service.AuditService;
 import com.bank.antifraud.service.SuspiciousAccountTransfersService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -26,12 +31,12 @@ public class SuspiciousAccountTransfersController {
     }
 
     @GetMapping("{id}")
-    public ResponseEntity<SuspiciousAccountTransfersEntity> getById(@PathVariable Long id) {
+    public ResponseEntity<SuspiciousAccountTransfersEntity> getById(@PathVariable BigInteger id) {
         return ResponseEntity.ok(suspiciousAccountTransfersService.findById(id));
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<Long> deleteAudit(@PathVariable Long id) {
+    public ResponseEntity<BigInteger> deleteAudit(@PathVariable BigInteger id) {
         suspiciousAccountTransfersService.delete(id);
         return ResponseEntity.ok(id);
     }
