+++ antifraud/src/main/java/com/bank/antifraud/controller/SuspiciousPhoneTransfersController.java
@@ -1,14 +1,19 @@
 package com.bank.antifraud.controller;
 
-import com.bank.antifraud.dto.AuditDto;
 import com.bank.antifraud.dto.SuspiciousPhoneTransfersDto;
-import com.bank.antifraud.entity.AuditEntity;
 import com.bank.antifraud.entity.SuspiciousPhoneTransfersEntity;
-import com.bank.antifraud.service.AuditService;
 import com.bank.antifraud.service.SuspiciousPhoneTransfersService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -26,12 +31,12 @@ public class SuspiciousPhoneTransfersController {
     }
 
     @GetMapping("{id}")
-    public ResponseEntity<SuspiciousPhoneTransfersEntity> getById(@PathVariable Long id) {
+    public ResponseEntity<SuspiciousPhoneTransfersEntity> getById(@PathVariable BigInteger id) {
         return ResponseEntity.ok(suspiciousPhoneTransfersService.findById(id));
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<Long> deleteAudit(@PathVariable Long id) {
+    public ResponseEntity<BigInteger> deleteAudit(@PathVariable BigInteger id) {
         suspiciousPhoneTransfersService.delete(id);
         return ResponseEntity.ok(id);
     }
