+++ antifraud/src/main/java/com/bank/antifraud/repository/SuspiciousAccountTransfersRepository.java
@@ -4,6 +4,8 @@ import com.bank.antifraud.entity.SuspiciousAccountTransfersEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface SuspiciousAccountTransfersRepository extends JpaRepository<SuspiciousAccountTransfersEntity, Long> {
+public interface SuspiciousAccountTransfersRepository extends JpaRepository<SuspiciousAccountTransfersEntity, BigInteger> {
 }
