+++ antifraud/src/main/java/com/bank/antifraud/service/AuditService.java
@@ -3,13 +3,13 @@ package com.bank.antifraud.service;
 import com.bank.antifraud.dto.AuditDto;
 import com.bank.antifraud.entity.AuditEntity;
 
+import java.math.BigInteger;
 import java.util.List;
-import java.util.Optional;
 
 public interface AuditService {
     public AuditEntity save(AuditDto auditDto);
-    public AuditEntity findById(Long id);
+    public AuditEntity findById(BigInteger id);
     public List<AuditEntity> findAll();
-    public void delete(Long id);
+    public void delete(BigInteger id);
     public AuditEntity update(AuditDto auditDto);
 }
